Got WasmFs e2e tested
